+++ src/constants/ASSETS_PATH.ts
@@ -1,0 +1,3 @@
+export const ASSETS_PATH = {
+  mainImage: './images/avengers-poster.jpg',
+};
