+++ src/constants/TIMER_CONFIG.ts
